+++ rse/main/parsers/base.py
@@ -11,6 +11,7 @@ with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 from rse.utils.file import read_file
 
 from datetime import datetime
+import copy
 import os
 import json
 import tempfile
@@ -18,7 +19,8 @@ import subprocess
 
 
 class Capturing:
-    """capture output from stdout and stderr into capture object.
+    """
+    capture output from stdout and stderr into capture object.
     This is based off of github.com/vsoch/gridtest but modified
     to write files. The stderr and stdout are set to temporary files at
     the init of the capture, and then they are closed when we exit. This
@@ -64,7 +66,8 @@ class Capturing:
 
     @property
     def err(self):
-        """Return error stream. Returns empty string if empty or doesn't exist.
+        """
+        Return error stream. Returns empty string if empty or doesn't exist.
         Returns (str) : error stream written to file
         """
         if os.path.exists(self.stderr.name):
@@ -78,12 +81,16 @@ class Capturing:
 
 
 class ParserBase:
-    """A parser base exists to extract and format repository metadata."""
+    """
+    A parser base exists to extract and format repository metadata.
+    """
 
     name = "base"
 
     def __init__(self, uid=None):
-        """set a unique id that includes parser name (type) and unique identifier)"""
+        """
+        set a unique id that includes parser name (type) and unique identifier)
+        """
         self.uid = None
         if uid is not None:
             self.set_uid(uid)
@@ -91,7 +98,9 @@ class ParserBase:
             self.data = {}
 
     def set_uid(self, uid):
-        """Given a unique resource identifier, set it for the parser"""
+        """
+        Given a unique resource identifier, set it for the parser
+        """
         uid = self._set_uid(uid)
         self.uid = "%s/%s" % (self.name, uid)
 
@@ -102,7 +111,8 @@ class ParserBase:
         raise NotImplementedError
 
     def load(self, data):
-        """If a repository has already been instantiated, we might want to load
+        """
+        If a repository has already been instantiated, we might want to load
         data into a parser to interact with it
         """
         if isinstance(data, str):
@@ -110,10 +120,19 @@ class ParserBase:
         self.data = data
 
     def _export_common(self):
-        """export common repo variables such as timestamp when it was updated.
-        This might include envars at some point, but we'd need to be careful.
         """
-        return {"timestamp": str(datetime.now())}
+        export common repo variables such as timestamp when it was updated,
+        and we try to add the common variables like description, etc.
+        """
+        res = {"timestamp": str(datetime.now())}
+        for field in ["description", "avatar", "url"]:
+            try:
+                # If we load again, "data" will be a field (the automated data from the parser)
+                func = getattr(self, f"get_{field}")
+                res[field] = func(self.data) or func(self.data.get("data"))
+            except:
+                pass
+        return res
 
     def get_url(self, data):
         """a common function for a parser to return the html url for the
@@ -122,26 +141,38 @@ class ParserBase:
         raise NotImplementedError
 
     def get_description(self, data):
-        """a common function for a parser to return a description."""
+        """
+        a common function for a parser to return a description.
+        """
         raise NotImplementedError
 
     def export(self):
-        """return data as json. This is intended to save to the software database.
+        """
+        return data as json. This is intended to save to the software database.
         Any important parser specific metadata should be added to self.data
         """
         # Get common context (e.g., pwd)
-        common = self._export_common()
-        common.update(self.data)
-        return common
+        data = copy.deepcopy(self.data)
+        for k, v in self._export_common().items():
+            if v:
+                data[k] = v
+        if "avatar" not in data:
+            import IPython
+
+            IPython.embed()
+            sys.exit()
+        return data
 
     def get_metadata(self, uri, **kwargs):
-        """The get_metadata function should take a general URI for a parser
+        """
+        The get_metadata function should take a general URI for a parser
         and populate the self.data
         """
         raise NotImplementedError
 
     def capture(self, cmd):
-        """capture is a helper function to capture a shell command. We
+        """
+        capture is a helper function to capture a shell command. We
         use Capturing and then save attributes like the pid, output, error
         to it, and return to the calling function. For example:
 
@@ -176,7 +207,8 @@ class ParserBase:
         return capture
 
     def get_setting(self, key, default=None):
-        """Get a setting, meaning that we first check the environment, then
+        """
+        Get a setting, meaning that we first check the environment, then
         the config file, and then (if provided) a default.
         """
         # First preference to environment