+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip.
 
 ## [0.0.x](https://github.com/rseng/rse/tree/master) (0.0.x)
+ - adding custom jekyll exporter for static site (0.0.38)
+ - import bug (0.0.37)
  - adding imperial college london research software directory (0.0.36)
  - adding molssi scraper (0.0.35)
  - adding ropensci scraper (0.0.34)